move formatQueryDetails to separate util file
